+++ api/modules/university/migrations/20170320135200-add-university-score.ts
@@ -1,7 +1,7 @@
 module.exports = {
     up: function(queryInterface, Sequelize) {
         return queryInterface.addColumn(
-            'university', 'score', Sequelize.ARRAY(Sequelize.INTEGER)
+            'university', 'score', Sequelize.ARRAY(Sequelize.FLOAT)
         );
     },
     down: function(queryInterface, Sequelize) {
