+++ api/modules/university/migrations/20170320153400-add-program-score.ts
@@ -1,7 +1,7 @@
 module.exports = {
     up: function(queryInterface, Sequelize) {
         return queryInterface.addColumn(
-            'program', 'score', Sequelize.ARRAY(Sequelize.INTEGER)
+            'program', 'score', Sequelize.ARRAY(Sequelize.FLOAT)
         );
     },
     down: function(queryInterface, Sequelize) {
